MAJOR: Refactor entire project to convert as much React to Astro as possible with the goal of reducing bundle size and improving performance
