+++ package.json
@@ -16,6 +16,7 @@
     "@astrojs/check": "^0.9.9",
     "@astrojs/react": "6.0.0-alpha.0",
     "@headlessui/react": "^2.2.10",
+    "@lucide/astro": "^1.17.0",
     "@radix-ui/react-dialog": "^1.1.16",
     "@radix-ui/react-visually-hidden": "^1.2.5",
     "@tanstack/react-store": "^0.11.0",
