+++ src/components/TypeIcon.astro
@@ -1,0 +1,57 @@
+---
+import { cva } from 'cva'
+
+import { TypeKey, TypeLabels } from '@/lib/utils/pokeapi-helpers'
+
+const variants = cva({
+  base: 'flex flex-row items-center justify-center',
+  variants: {
+    variant: {
+      [TypeKey.Normal]: 'bg-normal',
+      [TypeKey.Fighting]: 'bg-fighting',
+      [TypeKey.Flying]: 'bg-flying',
+      [TypeKey.Poison]: 'bg-poison',
+      [TypeKey.Ground]: 'bg-ground',
+      [TypeKey.Rock]: 'bg-rock',
+      [TypeKey.Bug]: 'bg-bug',
+      [TypeKey.Ghost]: 'bg-ghost',
+      [TypeKey.Steel]: 'bg-steel',
+      [TypeKey.Fire]: 'bg-fire',
+      [TypeKey.Water]: 'bg-water',
+      [TypeKey.Grass]: 'bg-grass',
+      [TypeKey.Electric]: 'bg-electric',
+      [TypeKey.Psychic]: 'bg-psychic',
+      [TypeKey.Ice]: 'bg-ice',
+      [TypeKey.Dragon]: 'bg-dragon',
+      [TypeKey.Dark]: 'bg-dark',
+      [TypeKey.Fairy]: 'bg-fairy',
+    },
+    size: {
+      small: 'size-5 rounded-xs',
+      medium: 'size-6 rounded-sm',
+      large: 'size-8 rounded-md',
+    },
+  },
+})
+
+interface Props {
+  variant: string
+  size?: 'small' | 'medium' | 'large'
+  class?: string
+}
+
+const { variant, size = 'large', class: className } = Astro.props
+const name = TypeLabels[variant as TypeKey]
+const imageUrl = `${variant}.png`
+const dimensions = { small: 20, medium: 28, large: 32 }
+---
+
+<div title={name} class:list={['relative', variants({ variant: variant as TypeKey, size }), className]}>
+  <img
+    src={imageUrl}
+    alt={name}
+    width={dimensions[size]}
+    height={dimensions[size]}
+    class="absolute object-contain"
+  />
+</div>
