+++ src/components/TypePill.astro
@@ -1,7 +1,7 @@
-import clsx from 'clsx/lite'
+---
 import { cva } from 'cva'
 
-import { TypeLabels, TypeKey } from '@/lib/utils/pokeapi-helpers'
+import { TypeKey, TypeLabels } from '@/lib/utils/pokeapi-helpers'
 
 const variants = cva({
   base: 'flex flex-row items-center',
@@ -34,36 +34,25 @@ const variants = cva({
   },
 })
 
-export default function TypePill({
-  variant,
-  size = 'large',
-  className,
-}: {
+interface Props {
   variant: string
   size?: 'small' | 'medium' | 'large'
-  className?: string
-}) {
-  const name = TypeLabels[variant as TypeKey]
-  const imageUrl = `${variant}.png`
-  const dimensions = {
-    small: 20,
-    medium: 24,
-    large: 28,
-  }
-
-  return (
-    <div className={clsx(variants({ size }), 'bg-zinc-200 dark:bg-zinc-800', className)}>
-      <img
-        src={imageUrl}
-        alt={name}
-        width={dimensions[size]}
-        height={dimensions[size]}
-        className={clsx(
-          variants({ variant: variant as TypeKey }),
-          'aspect-square rounded-xs object-contain'
-        )}
-      />
-      <p className="items-center text-black uppercase dark:text-white">{name}</p>
-    </div>
-  )
+  class?: string
 }
+
+const { variant, size = 'large', class: className } = Astro.props
+const name = TypeLabels[variant as TypeKey]
+const imageUrl = `${variant}.png`
+const dimensions = { small: 20, medium: 24, large: 28 }
+---
+
+<div class:list={[variants({ size }), 'bg-zinc-200 dark:bg-zinc-800', className]}>
+  <img
+    src={imageUrl}
+    alt={name}
+    width={dimensions[size]}
+    height={dimensions[size]}
+    class:list={[variants({ variant: variant as TypeKey }), 'aspect-square rounded-xs object-contain']}
+  />
+  <p class="items-center text-black uppercase dark:text-white">{name}</p>
+</div>
