--- src/components/details/MonsterMetadata.tsx
@@ -1,290 +1,0 @@
-import clsx from 'clsx/lite'
-import { Mars, Venus } from 'lucide-react'
-import type { Pokemon, PokemonSpecies, PokemonStat, EggGroup, GrowthRate } from 'pokedex-promise-v2'
-import { useId } from 'react'
-
-import { StatLabels, type StatKey, getTranslation } from '@/lib/utils/pokeapi-helpers'
-
-const calculateCatchProbability = (captureRate: number) => {
-  const a = captureRate / 3
-  const captureValue = 65535 / Math.pow(255 / a, 0.1875)
-  const captureProbability = Math.pow(captureValue / 65535, 4)
-  return Math.fround(captureProbability) * 100
-}
-
-const convertHeight = (heightInDecimeters: number) => ({
-  feet: Math.round(heightInDecimeters / 3.048).toFixed(0),
-  inches: (Math.round(heightInDecimeters * 3.937) % 12).toFixed(0),
-  meters: Math.fround(heightInDecimeters / 10).toFixed(1),
-})
-
-const convertWeight = (weightInHectograms: number) => ({
-  pounds: Math.fround(weightInHectograms / 4.536).toFixed(1),
-  kilograms: (weightInHectograms / 10).toFixed(1),
-})
-
-const calculateGenderRates = (genderRate: number) => ({
-  maleRate: Math.fround((1 - genderRate / 8) * 100).toFixed(1),
-  femaleRate: Math.fround((genderRate / 8) * 100).toFixed(1),
-})
-
-function MetadataCard({ title, children }: { title: string; children?: React.ReactNode }) {
-  return (
-    <div
-      className={clsx(
-        'flex flex-col gap-2',
-        'inset-ring-1 inset-ring-zinc-200 dark:inset-ring-zinc-800',
-        'rounded-xl p-4'
-      )}
-    >
-      <h2 className="mb-1 text-xs font-medium tracking-wide text-zinc-700 uppercase dark:text-zinc-300">
-        {title}
-      </h2>
-      <div className="flex flex-1 flex-col">{children}</div>
-    </div>
-  )
-}
-
-function ValueUnit({ value, unit }: { value: string | number; unit?: string }) {
-  return (
-    <p className="flex gap-x-1">
-      <span className="text-base font-light">{value}</span>
-      {unit && <span className="text-base text-zinc-600 dark:text-zinc-400">{unit}</span>}
-    </p>
-  )
-}
-
-function CatchRateMetadata({ catchRate }: { catchRate: number }) {
-  const probability = calculateCatchProbability(catchRate).toFixed(2)
-
-  return (
-    <MetadataCard title="Catch rate">
-      <ValueUnit value={catchRate} unit="/ 255" />
-      <ValueUnit value={probability} unit="%" />
-    </MetadataCard>
-  )
-}
-
-function EffortValueYieldMetadata({ stats }: { stats: PokemonStat[] }) {
-  const evStats = stats.filter((stat) => stat.effort > 0)
-
-  if (evStats.length === 0) {
-    return (
-      <MetadataCard title="EV yield">
-        <p>None</p>
-      </MetadataCard>
-    )
-  }
-
-  return (
-    <MetadataCard title="EV yield">
-      <ul className="flex flex-col">
-        {evStats.map((stat) => (
-          <li key={stat.stat.name} className="flex gap-x-1">
-            <ValueUnit value={stat.effort} unit={StatLabels[stat.stat.name as StatKey]} />
-          </li>
-        ))}
-      </ul>
-    </MetadataCard>
-  )
-}
-
-function SizeMetadata({ height, weight }: { height: number; weight: number }) {
-  const { feet, inches, meters } = convertHeight(height)
-  const { pounds, kilograms } = convertWeight(weight)
-
-  return (
-    <MetadataCard title="Size">
-      <div className="flex flex-col gap-2">
-        <div>
-          <span className="mb-1 block text-xs font-semibold text-zinc-500 dark:text-zinc-400">
-            Height
-          </span>
-          <span className="sr-only">
-            {feet} feet {inches} inches, or {meters} meters
-          </span>
-          <div aria-hidden="true">
-            <ValueUnit value={meters} unit="m" />
-          </div>
-        </div>
-        <div>
-          <span className="mb-1 block text-xs font-semibold text-zinc-500 dark:text-zinc-400">
-            Weight
-          </span>
-          <span className="sr-only">
-            {pounds} pounds, or {kilograms} kilograms
-          </span>
-          <div aria-hidden="true">
-            <ValueUnit value={kilograms} unit="kg" />
-          </div>
-        </div>
-      </div>
-    </MetadataCard>
-  )
-}
-
-function BreedingMetadata({
-  hatchCounter,
-  eggGroups,
-}: {
-  hatchCounter: number | null
-  eggGroups: EggGroup[]
-}) {
-  return (
-    <MetadataCard title="Breeding">
-      <div className="flex flex-col gap-2">
-        <div>
-          <span className="mb-1 block text-xs font-semibold text-zinc-500 dark:text-zinc-400">
-            Egg Groups
-          </span>
-          <div className="flex flex-wrap gap-1">
-            {eggGroups.length > 0 ? (
-              eggGroups.map((group, i) => (
-                <span key={group.name} className="font-medium">
-                  {getTranslation(group.names, 'name')}
-                  {i < eggGroups.length - 1 && ','}
-                </span>
-              ))
-            ) : (
-              <span>—</span>
-            )}
-          </div>
-        </div>
-        <div>
-          <span className="mb-1 block text-xs font-semibold text-zinc-500 dark:text-zinc-400">
-            Hatch Cycles
-          </span>
-          <div>
-            <span>{hatchCounter ? hatchCounter.toLocaleString() : '—'}</span>
-          </div>
-        </div>
-      </div>
-    </MetadataCard>
-  )
-}
-
-function GenderRatioDonut({ male, female }: { male: number; female: number }) {
-  const gradientIdPrefix = useId().replace(/:/g, '')
-  const maleGradientId = `${gradientIdPrefix}-male-gradient`
-  const femaleGradientId = `${gradientIdPrefix}-female-gradient`
-  const width = 128
-  const height = 128
-  const radius = 56
-  const strokeWidth = 6
-  const circumference = 2 * Math.PI * radius
-  const maleArc = Number(((male / 100) * circumference).toFixed(6))
-  const femaleArc = Number(((female / 100) * circumference).toFixed(6))
-  const maleDashArray = `${maleArc} ${Number((circumference - maleArc).toFixed(6))}`
-  const femaleDashArray = `${femaleArc} ${Number((circumference - femaleArc).toFixed(6))}`
-
-  return (
-    <div className="relative">
-      <svg width={width} height={height} aria-label="Gender ratio donut chart" role="img">
-        <defs>
-          <linearGradient id={maleGradientId} x1="0%" y1="0%" x2="100%" y2="100%">
-            <stop offset="0%" stopColor="#93c5fd" />
-            <stop offset="100%" stopColor="#2563eb" />
-          </linearGradient>
-          <linearGradient id={femaleGradientId} x1="0%" y1="0%" x2="100%" y2="100%">
-            <stop offset="0%" stopColor="#db2777" />
-            <stop offset="100%" stopColor="#f9a8d4" />
-          </linearGradient>
-        </defs>
-        <g transform={`translate(${width / 2} ${height / 2}) rotate(-90)`}>
-          {maleArc > 0 && (
-            <circle
-              r={radius}
-              cx={0}
-              cy={0}
-              fill="none"
-              strokeWidth={strokeWidth}
-              strokeDasharray={maleDashArray}
-              stroke={`url(#${maleGradientId})`}
-            />
-          )}
-          {femaleArc > 0 && (
-            <circle
-              r={radius}
-              cx={0}
-              cy={0}
-              fill="none"
-              strokeWidth={strokeWidth}
-              strokeDasharray={femaleDashArray}
-              strokeDashoffset={-maleArc}
-              stroke={`url(#${femaleGradientId})`}
-            />
-          )}
-        </g>
-      </svg>
-      <div className="absolute top-1/2 left-1/2 flex size-full -translate-x-1/2 -translate-y-1/2 flex-col items-center justify-center">
-        <div className="flex gap-x-1">
-          <span className="flex flex-row items-center gap-x-1 text-blue-800 dark:text-blue-200">
-            <Mars size={20} />
-            <ValueUnit value={male} unit="%" />
-          </span>
-        </div>
-        <div className="flex gap-x-1">
-          <span className="flex flex-row items-center gap-x-1 text-base font-light text-pink-800 dark:text-pink-200">
-            <Venus size={20} />
-            <ValueUnit value={female} unit="%" />
-          </span>
-        </div>
-      </div>
-    </div>
-  )
-}
-
-function GenderRatioMetadata({ genderRate }: { genderRate: number }) {
-  if (genderRate === -1) {
-    return (
-      <MetadataCard title="Gender ratio">
-        <p>Genderless</p>
-      </MetadataCard>
-    )
-  }
-
-  const { femaleRate, maleRate } = calculateGenderRates(genderRate)
-
-  return (
-    <MetadataCard title="Gender ratio">
-      <div className="flex items-center">
-        <GenderRatioDonut female={parseFloat(femaleRate)} male={parseFloat(maleRate)} />
-      </div>
-    </MetadataCard>
-  )
-}
-
-function GrowthRateMetadata({ growthRate }: { growthRate: GrowthRate }) {
-  const description = getTranslation(growthRate.descriptions, 'description')!
-  const maxExperience = growthRate.levels.find((level) => level.level === 100)!.experience
-
-  return (
-    <MetadataCard title="Growth rate">
-      <ValueUnit value={maxExperience.toLocaleString()} unit="exp" />
-      <ValueUnit value={description} />
-    </MetadataCard>
-  )
-}
-
-export default function MonsterMetadata({
-  species,
-  pokemon,
-  eggGroups,
-  growthRate,
-}: {
-  species: PokemonSpecies
-  pokemon: Pokemon
-  eggGroups: EggGroup[]
-  growthRate: GrowthRate
-}) {
-  return (
-    <>
-      <SizeMetadata height={pokemon.height} weight={pokemon.weight} />
-      <GenderRatioMetadata genderRate={species.gender_rate} />
-      <CatchRateMetadata catchRate={species.capture_rate} />
-      <BreedingMetadata hatchCounter={species.hatch_counter} eggGroups={eggGroups} />
-      <GrowthRateMetadata growthRate={growthRate} />
-      <EffortValueYieldMetadata stats={pokemon.stats} />
-    </>
-  )
-}
