+++ src/components/details/cosmetics/CosmeticsSection.astro
@@ -1,0 +1,46 @@
+---
+import type { Pokemon, PokemonForm } from 'pokedex-promise-v2'
+
+import { getTranslation } from '@/lib/utils/pokeapi-helpers'
+
+interface Props {
+  pokemon: Pokemon
+  forms: PokemonForm[]
+}
+
+const { pokemon, forms } = Astro.props
+const cosmeticForms = forms.filter((form) => form.name !== pokemon.name)
+const imageId = pokemon.id.toString().padStart(4, '0')
+---
+
+<section
+  class="@container flex flex-col gap-4 rounded-xl p-4 inset-ring-1 inset-ring-zinc-200 dark:inset-ring-zinc-800"
+>
+  <h2 class="text-xl font-medium text-black dark:text-white">Cosmetic forms</h2>
+  {cosmeticForms.length === 0 ? (
+    <p class="flex items-baseline gap-2">
+      <span class="text-lg text-pretty text-zinc-700 dark:text-zinc-300">
+        No cosmetic forms available for this Pokémon.
+      </span>
+    </p>
+  ) : (
+    <ul
+      class="grid grid-cols-1 gap-4 @sm:grid-cols-2 @lg:grid-cols-3 @xl:grid-cols-4 @4xl:grid-cols-6"
+    >
+      {cosmeticForms.map((form) => {
+        const name = getTranslation(form.form_names, 'name')!
+        const imageUrl = form.is_default
+          ? `https://raw.githubusercontent.com/blai30/PokemonSpritesDump/refs/heads/main/sprites/sprite_${imageId}_s0.webp`
+          : `https://raw.githubusercontent.com/blai30/PokemonSpritesDump/refs/heads/main/sprites/sprite_${imageId}_${form.name}_s0.webp`
+        return (
+          <li
+            class="flex flex-col items-center gap-2 rounded-lg bg-linear-to-br from-white to-zinc-100 p-4 dark:from-zinc-900 dark:to-zinc-950"
+          >
+            <span title={name} class="text-sm">{name}</span>
+            <img src={imageUrl} alt={name} width={128} height={128} class="object-contain" />
+          </li>
+        )
+      })}
+    </ul>
+  )}
+</section>
