+++ src/components/details/localization/LocalizationSection.astro
@@ -1,0 +1,30 @@
+---
+import type { PokemonSpecies } from 'pokedex-promise-v2'
+
+interface Props {
+  species: PokemonSpecies
+}
+
+const { species } = Astro.props
+const sortedNames = species.names.toSorted((a, b) => a.language.name.localeCompare(b.language.name))
+---
+
+<section
+  class="@container flex flex-col gap-4 rounded-xl p-4 inset-ring-1 inset-ring-zinc-200 dark:inset-ring-zinc-800"
+>
+  <h2 class="text-xl font-medium text-black dark:text-white">Localization</h2>
+  <ul
+    class="grid max-w-lg grid-flow-col grid-cols-1 grid-rows-11 gap-2 @md:grid-cols-2 @md:grid-rows-6"
+  >
+    {sortedNames.map((name) => (
+      <li class="flex w-full flex-col items-center">
+        <p class="flex w-full items-center gap-3">
+          <span title={name.language.name} class="w-1/3 text-right font-medium">
+            {name.language.name}
+          </span>
+          <span class="text-zinc-700 dark:text-zinc-300">{name.name}</span>
+        </p>
+      </li>
+    ))}
+  </ul>
+</section>
